% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm-helpers.R
\name{calcfrex}
\alias{calcfrex}
\alias{calclift}
\alias{calcscore}
\title{stm-compatible label scorers (FREX / lift / score)}
\usage{
calcfrex(logbeta, w = 0.5, wordcounts = NULL)

calclift(logbeta, wordcounts)

calcscore(logbeta)
}
\arguments{
\item{logbeta}{K x V log topic-word matrix.}

\item{w}{FREX frequency/exclusivity weight.}

\item{wordcounts}{Corpus term frequencies (enables the James-Stein shrinkage).}
}
\value{
A V x K matrix of ordered word indices.
}
\description{
Ports of \code{stm:::calcfrex}/\code{calclift}/\code{calcscore}. Each takes a K x V \code{logbeta}
(log topic-word matrix) and returns a V x K matrix whose columns are the word
indices ordered most- to least-characteristic for each topic.
}
