% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faSTM-package.R
\docType{package}
\name{faSTM-package}
\alias{faSTM}
\alias{faSTM-package}
\title{faSTM: fast structural topic models with an stm-compatible object}
\description{
faSTM fits the logistic-normal structural topic model (prevalence + content
covariates) using the Rust core from \code{topica}, and returns an object
compatible with the \code{stm} package. Text preparation
(\code{\link[stm:textProcessor]{stm::textProcessor()}}, \code{\link[stm:prepDocuments]{stm::prepDocuments()}}) and post-fit tools
(\code{\link[stm:labelTopics]{stm::labelTopics()}}, \code{\link[stm:plot.STM]{stm::plot.STM()}}, \code{\link[stm:findThoughts]{stm::findThoughts()}},
\code{\link[stm:sageLabels]{stm::sageLabels()}}, \code{\link[stm:toLDAvis]{stm::toLDAvis()}}) are reused from \code{stm} unchanged;
faSTM contributes the fast fit (\code{\link[=stm]{stm()}}) and an honest \code{\link[=estimateEffect]{estimateEffect()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nealcaren.github.io/faSTM/}
  \item \url{https://github.com/nealcaren/faSTM}
  \item Report bugs at \url{https://github.com/nealcaren/faSTM/issues}
}

}
\author{
\strong{Maintainer}: Neal Caren \email{neal.caren@unc.edu}

Authors:
\itemize{
  \item Neal Caren \email{neal.caren@unc.edu}
}

}
\keyword{internal}
