% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{glance.faSTM}
\alias{glance.faSTM}
\title{One-row model summary for a faSTM fit}
\usage{
\method{glance}{faSTM}(x, ...)
}
\arguments{
\item{x}{A faSTM fit.}

\item{...}{Unused.}
}
\value{
A one-row data.frame.
}
\description{
One-row model summary for a faSTM fit
}
