% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm-helpers.R
\name{makeDesignMatrix}
\alias{makeDesignMatrix}
\title{Build a (sparse) design matrix for new data (stm-compatible)}
\usage{
makeDesignMatrix(formula, origData, newData, sparse = TRUE, ...)
}
\arguments{
\item{formula}{A model formula.}

\item{origData}{Data defining the terms/levels.}

\item{newData}{Data to build the matrix for.}

\item{sparse}{Return a sparse matrix.}

\item{...}{Ignored.}
}
\value{
A (sparse) design matrix.
}
\description{
Port of \code{stm:::makeDesignMatrix}: builds the model matrix for \code{newData} using
the term structure and factor levels of \code{origData}.
}
