% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{predict.faSTM}
\alias{predict.faSTM}
\title{Predict topic proportions for new documents}
\usage{
\method{predict}{faSTM}(object, newdata, ...)
}
\arguments{
\item{object}{A faSTM fit.}

\item{newdata}{New documents (corpus / dfm / matrix / stm-style list).}

\item{...}{Passed to \code{\link[=fit_new_documents]{fit_new_documents()}}.}
}
\value{
A new-documents x K matrix of topic proportions.
}
\description{
Predict topic proportions for new documents
}
